% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all.R
\name{get.quantiles}
\alias{get.quantiles}
\title{Extract quantile estimates from an LMS fit}
\usage{
get.quantiles(lmsobj, targetx, pvec = c(0.05, 0.1, 0.25, 0.5, 0.75, 0.9, 0.95))
}
\arguments{
\item{lmsobj}{an object of class \code{"lmsqreg.fit"}}

\item{targetx}{numeric vector of x values at which to evaluate quantiles}

\item{pvec}{numeric vector of probabilities}
}
\value{
An object of class \code{"qsys.out"} with components \code{outmat}
(matrix of quantile values, rows = quantiles, cols = \code{targetx}),
\code{targetx}, \code{pcts}, and \code{edf}.
}
\description{
Replicates the internal \code{qsys()} computation outside of \code{lmsqreg.fit()},
allowing quantiles to be evaluated at a user-supplied grid of x values.
}
